+++ regression_tests/web/__init__.py
@@ -1,8 +1,0 @@
-"""
-    Web interface for regression tests.
-"""
-
-from flask import Flask
-app = Flask(__name__)
-
-from regression_tests.web.views import *