+++ runner.py
@@ -19,22 +19,14 @@ from datetime import datetime
 from regression_tests.clang import setup_clang_bindings
 from regression_tests.cmd_runner import CmdRunner
 from regression_tests.config import parse_standard_config_files
-from regression_tests.db import DB
-from regression_tests.email import prepare_email
-from regression_tests.email import send_email
 from regression_tests.filesystem.directory import Directory
-from regression_tests.git import Repository
-from regression_tests.git import add_username
 from regression_tests.io import print_error
 from regression_tests.io import print_prologue
 from regression_tests.io import print_summary
 from regression_tests.io import print_test_results
 from regression_tests.logging import setup_logging
-from regression_tests.qos import WithQoS
-from regression_tests.retdec_builder import build_retdec
 from regression_tests.test_finder import find_tests
 from regression_tests.test_finder import get_tests_dir
-from regression_tests.test_results import NoTestResults
 from regression_tests.test_results import TestResults
 from regression_tests.test_results import TestsResults
 from regression_tests.test_settings import TestSettings
@@ -45,23 +37,14 @@ def parse_args():
     parser = argparse.ArgumentParser(description='Runner of regression tests.')
     parser.add_argument('tests_dir', type=str, metavar='PATH', nargs='?',
                         help='Explicit path to the tests.')
-    parser.add_argument('-b', '--build', action='store_true', dest='build',
-                        help='Build RetDec before testing (enabled if -c/--commit is used).')
-    parser.add_argument('-c', '--commit', type=str, metavar='SHA', dest='commit',
-                        help='Force commit to be tested.')
     parser.add_argument('-r', '--regexp', type=str, metavar='REGEXP', dest='regexp',
                         help='Run only tests matching the given regular expression.')
-    parser.add_argument('-R', '--resume', action='store_true', dest='resume',
-                        help='Do not run tests that have already run for the commit.')
     parser.add_argument('-t', '--tool', type=str, metavar='TOOL', dest='tool',
                         help='Run only tests for the given tool.')
     parser.add_argument('--skip-c-compilation-tests', action='store_true',
                         dest='skip_c_compilation_tests',
                         help='Skip tests that compile the output C source code.')
     args = parser.parse_args()
-
-    if args.commit is not None:
-        args.build = True
 
     return args
 
@@ -83,21 +66,6 @@ def ensure_all_required_settings_are_set(config):
                     "does not seem to point to Clang")
         sys.exit(1)
 
-    # [runner] -> retdec_build_dir
-    retdec_build_dir = config['runner']['retdec_build_dir']
-    if not retdec_build_dir:
-        print_error("no 'retdec_build_dir' in the [runner] section of config_local.ini "
-                    "(you have to add it)")
-        sys.exit(1)
-    elif not os.path.exists(retdec_build_dir):
-        print_error("'retdec_build_dir' in the [runner] section of config_local.ini "
-                    "points to a non-existing directory")
-        sys.exit(1)
-    elif not os.path.exists(os.path.join(retdec_build_dir, 'CMakeCache.txt')):
-        print_error("'retdec_build_dir' in the [runner] section of config_local.ini "
-                    "does not seem to point to a RetDec build directory")
-        sys.exit(1)
-
     # [runner] -> retdec_install_dir
     retdec_install_dir = config['runner']['retdec_install_dir']
     if not retdec_install_dir:
@@ -111,21 +79,6 @@ def ensure_all_required_settings_are_set(config):
     elif not os.path.exists(os.path.join(retdec_install_dir, 'bin', 'retdec-decompiler.py')):
         print_error("'retdec_install_dir' in the [runner] section of config_local.ini "
                     "does not seem to point to RetDec")
-        sys.exit(1)
-
-    # [runner] -> retdec_repo_dir
-    retdec_repo_dir = config['runner']['retdec_repo_dir']
-    if not retdec_repo_dir:
-        print_error("no 'retdec_repo_dir' in the [runner] section of config_local.ini "
-                    "(you have to add it)")
-        sys.exit(1)
-    elif not os.path.exists(retdec_repo_dir):
-        print_error("'retdec_repo_dir' in the [runner] section of config_local.ini "
-                    "points to a non-existing directory")
-        sys.exit(1)
-    elif not os.path.exists(os.path.join(retdec_repo_dir, 'src', 'bin2llvmir')):
-        print_error("'retdec_repo_dir' in the [runner] section of config_local.ini "
-                    "does not seem to point to a RetDec repository")
         sys.exit(1)
 
     # [runner] -> tests_root_dir
@@ -249,41 +202,6 @@ def should_skip_c_compilation_tests(config, args):
         config['runner'].getboolean('skip_c_compilation_tests')
 
 
-def run_retdec_build(config, db, cmd_runner, tested_commit):
-    """Runs a build of RetDec.
-
-    If the build fails, it prints and error message and sends an email to the
-    author of the tested commit (if needed).
-    """
-    build_start_date = datetime.now()
-    build_id = db.insert_build_started_info(
-        tested_commit,
-        build_start_date
-    )
-    build_info = build_retdec(
-        Directory(config['runner']['retdec_build_dir']),
-        cmd_runner,
-        int(config['runner']['build_procs'])
-    )
-
-    # Ensure that the same date appears after the build is finished
-    # (otherwise, there may be a difference of a few seconds).
-    build_info.start_date = build_start_date
-    db.insert_build_ended_info(build_id, build_info)
-
-    if not build_info.succeeded:
-        LINES_COUNT = 30
-        error_msg = 'failed to build RetDec; ' +\
-            'showing the last {} lines of the log:\n...\n'.format(LINES_COUNT) +\
-            '\n'.join(build_info.log.split('\n')[-LINES_COUNT:]).strip()
-        logging.error(error_msg)
-        print_error(error_msg)
-
-        send_email_for_commit_if_needed(config, db, tested_commit)
-
-        sys.exit(1)
-
-
 def get_num_of_procs_for_tests(config):
     """Returns the number of processes to be used to run the tests."""
     tests_procs = int(config['runner']['tests_procs'])
@@ -336,8 +254,7 @@ def initialize_worker(mp_lock):
     # The lock has to be made global through an initialization function when
     # creating mp.Pool(). Otherwise, interpreter instances on Windows would not
     # share a single lock but each interpreter would create its own lock. This
-    # would cause race conditions when emitting output or accessing the
-    # database.
+    # would cause race conditions when emitting output.
     # Based on http://stackoverflow.com/a/28721419.
     global lock
     lock = mp_lock
@@ -361,9 +278,6 @@ def ordered_indexes(test_cases):
     #
     # We use a heuristic approach by running first tests that are known to be
     # long-running. Then, we run other tests.
-    #
-    # An alternative (and better) approach would be to consult the database to
-    # find out the precise running time of tests from the previous commit.
 
     def format_num_prefix(n):
         return str(n).zfill(2)
@@ -388,7 +302,7 @@ def ordered_indexes(test_cases):
     return sorted(range(len(test_cases)), key=test_case_key)
 
 
-def run_test_cases(test_cases, tested_commit, procs, resume, lock):
+def run_test_cases(test_cases, procs, lock):
     """Runs the given test cases and returns a list of results."""
     pool = mp.Pool(
         processes=procs,
@@ -407,11 +321,9 @@ def run_test_cases(test_cases, tested_commit, procs, resume, lock):
     signal.signal(signal.SIGTERM, handler)
 
     try:
-        run_test_case_on_index_args = [
-            (i, tested_commit, resume) for i in ordered_indexes(test_cases)
-        ]
+        run_test_case_on_index_args = ordered_indexes(test_cases)
         test_results = TestsResults(
-            pool.starmap(
+            pool.map(
                 run_test_case_on_index,
                 run_test_case_on_index_args,
                 # Send tasks to processes one by one instead of sending them a
@@ -430,36 +342,22 @@ def run_test_cases(test_cases, tested_commit, procs, resume, lock):
         pool.join()
 
 
-def run_test_case_on_index(i, tested_commit, with_resume):
-    """Runs a test case on the given index for the given tested commit."""
+def run_test_case_on_index(i):
+    """Runs a test case on the given index."""
     global cmd_runner
-    global db
     global lock
     global test_cases
     global tools_dir
 
     test_case = test_cases[i]
 
-    # Check whether the test case should run. We only need to consult the
-    # database if resume is requested, i.e. when only running tests that have
-    # not run yet.
-    test_case_should_run = True
-    if with_resume:
-        with lock:
-            test_case_should_run = not db.has_test_run_for_commit(
-                test_case.module_name, test_case.name, tested_commit)
-
-    if test_case_should_run:
-        tool_runner = test_case.test_settings.get_tool_runner(
-            cmd_runner,
-            tools_dir
-        )
-        test_results = run_test_case(test_case, tool_runner)
-        with lock:
-            db.insert_test_results(test_results, tested_commit)
-            print_test_results(test_results)
-    else:
-        test_results = NoTestResults(test_case.module_name, test_case.name)
+    tool_runner = test_case.test_settings.get_tool_runner(
+        cmd_runner,
+        tools_dir
+    )
+    test_results = run_test_case(test_case, tool_runner)
+    with lock:
+        print_test_results(test_results)
     return test_results
 
 
@@ -510,67 +408,6 @@ def run_test_case(test_case, tool_runner):
     )
 
 
-def send_email_for_commit_if_needed(config, db, commit):
-    """Sends an email for the given commit, depending on several conditions.
-    """
-    if not config['email'].getboolean('enabled'):
-        # Sending of emails is disabled.
-        return
-
-    if db.email_sent_for_commit(commit):
-        # The email has already been sent.
-        return
-
-    commit_results = db.get_results_for_commit(commit)
-    if (not commit_results.build_has_failed() and
-            not commit_results.has_failed_tests()):
-        # There is nothing to report.
-        return
-
-    if db.topmost_commit_has_succeeded():
-        # There is no need to send the email if the topmost commit has
-        # succeeded.
-        return
-
-    email = prepare_email(
-        commit_results,
-        config['email']['subject_prefix'],
-        config['web']['main_page_url'],
-        config['web']['wiki_page_url']
-    )
-
-    # Set proper addresses.
-    def set_addr_if_defined(addr_type):
-        addr_name = '{}_addr'.format(addr_type)
-        addr = config['email'][addr_name]
-        if addr:
-            setattr(email, addr_name, addr)
-    set_addr_if_defined('from')
-    set_addr_if_defined('to')
-    set_addr_if_defined('reply_to')
-    set_addr_if_defined('cc')
-
-    # Send the email.
-    logging.info("sending email to '{}' concerning commit {}".format(
-        email.to_addr, commit.short_hash()))
-    send_email(
-        email,
-        config['email']['smtp_server'],
-        config['email']['smtp_port'],
-        config['email']['smtp_user'],
-        config['email']['smtp_pass']
-    )
-    db.insert_email_for_commit(email, commit)
-
-
-def request_username_for_repo(repo_url):
-    """Requests a username for the given repository and returns a new URL,
-    containing the username.
-    """
-    username = input('Enter username for {}: '.format(repo_url))
-    return add_username(username.strip(), repo_url)
-
-
 try:
     # Config.
     config = parse_standard_config_files()
@@ -588,16 +425,6 @@ try:
     # Arguments.
     args = parse_args()
 
-    # Quality of Service.
-    qos_enabled = config.getboolean('qos', 'enabled')
-    qos_max_tries = int(config['qos']['max_tries'])
-    qos_wait_time = int(config['qos']['wait_time'])
-
-    # Database.
-    db = DB(config['db']['conn_url'])
-    if qos_enabled:
-        db = WithQoS(db, qos_max_tries, qos_wait_time)
-
     # Command runner.
     cmd_runner = CmdRunner()
     tools_dir = Directory(os.path.join(config['runner']['retdec_install_dir'], 'bin'))
@@ -607,8 +434,7 @@ try:
 
     # Tests.
 
-    # We need a lock due to parallel access (1) to the database and (2)
-    # to the filesystem to prevent mixed text on the standard output.
+    # We need a lock due to prevent mixed text on the standard output.
     lock = mp.Lock()
 
     # Directories.
@@ -618,8 +444,6 @@ try:
 
     # Workers (spawned processes by the multiprocessing module) need access to
     # the test cases, so get them. They need to be stored in a global variable.
-    # The main process gets them after a checkout of the appropriate commit
-    # from the repository.
     if __name__ != '__main__':
         test_cases = get_test_cases_to_run(
             tests_dir,
@@ -631,65 +455,34 @@ try:
 
     if __name__ == '__main__':
         # The main process.
+        remove_results_from_previous_test_runs(tests_dir)
 
-        # Repository.
-        repo = Repository(config['runner']['retdec_repo_dir'], cmd_runner)
-        if qos_enabled:
-            repo = WithQoS(repo, qos_max_tries, qos_wait_time)
-
-        # Tested commit.
-        if args.commit is not None:
-            repo.update()
-            tested_commit = repo.get_commit_from_hash(args.commit)
-        else:
-            tested_commit = repo.get_current_commit()
-        db.initialize_commit_records(
-            tested_commit,
-            # When the user does not want to re-run tests that have already run
-            # (i.e. he or she wants to resume the execution), do not remove
-            # existing test results and sent emails from the database.
-            remove_test_results=not args.resume,
-            remove_emails=not args.resume
+        # Find tests.
+        test_cases = get_test_cases_to_run(
+            tests_dir,
+            tests_root_dir,
+            excluded_dirs,
+            config,
+            args
         )
-
-        with repo.checkout(tested_commit):
-            # RetDec build.
-            if args.build:
-                run_retdec_build(config, db, cmd_runner, tested_commit)
-
-            remove_results_from_previous_test_runs(tests_dir)
-
-            # Find tests.
-            test_cases = get_test_cases_to_run(
-                tests_dir,
-                tests_root_dir,
-                excluded_dirs,
-                config,
-                args
+        if not test_cases:
+            relative_excluded_dirs = ', '.join(
+                os.path.relpath(dir.path, tests_root_dir.path) for dir in excluded_dirs
             )
-            if not test_cases:
-                relative_excluded_dirs = ', '.join(
-                    os.path.relpath(dir.path, tests_root_dir.path) for dir in excluded_dirs
-                )
-                print_error('no tests found in {} (excluded directories: {})'.format(
-                    tests_dir.path,
-                    relative_excluded_dirs,
-                ))
-                sys.exit(1)
+            print_error('no tests found in {} (excluded directories: {})'.format(
+                tests_dir.path,
+                relative_excluded_dirs,
+            ))
+            sys.exit(1)
 
-            # Run them.
-            print_prologue(tests_dir.path, test_cases, tested_commit, args.resume)
-            tests_results = run_test_cases(
-                test_cases,
-                tested_commit,
-                procs=get_num_of_procs_for_tests(config),
-                resume=args.resume,
-                lock=lock
-            )
-            print_summary(tests_results)
-
-            # Send notifications (if needed).
-            send_email_for_commit_if_needed(config, db, tested_commit)
+        # Run them.
+        print_prologue(tests_dir.path, test_cases)
+        tests_results = run_test_cases(
+            test_cases,
+            procs=get_num_of_procs_for_tests(config),
+            lock=lock
+        )
+        print_summary(tests_results)
 
         sys.exit(0 if tests_results.succeeded else 1)
 except Exception: