+++ tests/web/__init__.py
@@ -1,3 +1,0 @@
-"""
-    Tests for the :mod:`regression_tests.web` package.
-"""